+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Adds
 
 * Allow to disable shortcut by setting the option `shortcut: false`
+* Adds a new color picker tool for the rich-text-widget toolbar that matches the existing `color` schema field. This also adds the same `colorPicker` and `format` options to the rich-text-widget confirmation that exist in the `color` schema field.
+* Add missing UI translation keys.
 
 ## 4.4.3 (2024-06-17)